# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Discovery Assistant API
  version: 1.0.0
  description: API for event, business, and offer discovery with conversational AI support
servers:
  - url: https://api.yourdomain.com/v1
paths:
...
  /conversations:
    post:
      summary: Start a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'

    get:
      summary: List all conversations
      responses:
        '200':
          description: Array of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
  /conversations/{conversationId}/messages:
    post:
      summary: Send a message and get system response
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sender
                - content
              properties:
                sender:
                  type: string
                  enum: [user, system]
                content:
                  type: string
      responses:
        '201':
          description: System response with suggested items
          content:
            application/json:
              schema:
                type: object
                properties:
                  system_message:
                    $ref: '#/components/schemas/Message'
                  suggested_items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SuggestedItem'
...
components:
  schemas:
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        context_summary:
          type: string
    Message:
      type: object
      properties:
        id:
          type: integer
        conversation_id:
          type: integer
        sender:
          type: string
          enum: [user, system]
        content:
          type: string
        message_type:
          type: string
          enum: [text, image, audio, video, interactive]
        created_at:
          type: string
          format: date-time
    SuggestedItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [event, business, offer]
        item_id:
          type: integer
        title:
          type: string
        description:
          type: string
        link:
          type: string